CXSPA-8773: move files from account to orgnization
